+++ src/main/java/gov/epa/bencloud/server/database/jooq/data/tables/VariableValue.java
@@ -65,7 +65,7 @@ public class VariableValue extends TableImpl<VariableValueRecord> {
     /**
      * The column <code>data.variable_value.grid_cell_id</code>.
      */
-    public final TableField<VariableValueRecord, Integer> GRID_CELL_ID = createField(DSL.name("grid_cell_id"), SQLDataType.INTEGER, this, "");
+    public final TableField<VariableValueRecord, Long> GRID_CELL_ID = createField(DSL.name("grid_cell_id"), SQLDataType.BIGINT, this, "");
 
     private VariableValue(Name alias, Table<VariableValueRecord> aliased) {
         this(alias, aliased, null);
@@ -136,7 +136,7 @@ public class VariableValue extends TableImpl<VariableValueRecord> {
     // -------------------------------------------------------------------------
 
     @Override
-    public Row5<Integer, Integer, Integer, Double, Integer> fieldsRow() {
+    public Row5<Integer, Integer, Integer, Double, Long> fieldsRow() {
         return (Row5) super.fieldsRow();
     }
 }
